+++ app/src/androidTest/java/io/dotanuki/demos/norris/test/ChipGroupContentAssertion.kt
@@ -1,4 +1,4 @@
-package io.dotanuki.demos.norris
+package io.dotanuki.demos.norris.test
 
 import android.view.View
 import androidx.core.view.children
