+++ features/facts/src/androidTest/java/io/dotanuki/norris/facts/test/FactsScreenshotTests.kt
@@ -1,4 +1,4 @@
-package io.dotanuki.norris.facts
+package io.dotanuki.norris.facts.test
 
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
