+++ features/search/src/androidTest/java/io/dotanuki/norris/search/test/SearchScreenshotTests.kt
@@ -1,4 +1,4 @@
-package io.dotanuki.norris.search
+package io.dotanuki.norris.search.test
 
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
